requirements.txt
beautifulsoup4==4.13.3
clickhouse_connect==0.8.15
fitz==0.0.1.dev2
fuzzywuzzy==0.18.0
gspread==6.1.4
numpy==2.2.2
office365==0.3.15
Office365_REST_Python_Client==2.5.11
openpyxl==3.1.5
pandas==2.2.3
protobuf==5.29.3
psutil==6.1.1
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyshorteners==1.0.1
python-dotenv==1.0.1
python_pptx==1.0.2
pywin32==308
Requests==2.32.3
selenium==4.28.1
SQLAlchemy==2.0.31
sqlalchemy_schemadisplay==2.0
statsmodels==0.14.4
Unidecode==1.3.8
webdriver_manager==4.0.1
